Seventh unit test baby steps.

+++ app/src/main/java/nl/visser/joram/mathapp/calculationModule/CalculatorForTwoValues.java
@@ -2,7 +2,7 @@ package nl.visser.joram.mathapp.calculationModule;
 
 public class CalculatorForTwoValues {
 
-    public static int calculateProduct(int valueOne, Operator operator, int valueTwo) {
+    public static double calculateProduct(double valueOne, Operator operator, double valueTwo) {
         if(Operator.PLUS.equals(operator)) {
             valueOne += valueTwo;
         } else if(Operator.MINUS.equals(operator)) {
